Updated cpu core.

+++ source/Gui.h
@@ -24,10 +24,10 @@ void uiOptions(void);
 void uiController(void);
 void uiDisplay(void);
 
-void debugIOUnmappedR(u8 port, u8 val);
-void debugIOUnmappedW(u8 port, u8 val);
-void debugIOUnimplR(u8 port, u8 val);
-void debugIOUnimplW(u8 port, u8 val);
+void debugIOUnmappedR(u16 port, u8 val);
+void debugIOUnmappedW(u16 port, u8 val);
+void debugIOUnimplR(u16 port, u8 val);
+void debugIOUnimplW(u16 port, u8 val);
 void debugDivideError(void);
 void debugUndefinedInstruction(void);
 void debugCrashInstruction(void);
